Added port 'playSequenceStarted'
This is an optional subscription from an elm client which comes as a simple boolean response to a request to play a note sequence.  The reason why it's important is that any sort of response to an elm request allows elm then to perform a further task if it wants to.  Without it, there is no mechanism to do 'andThen' after submitting a command.

In this case, the elm program may want (for example) to suspend the UI whilst the note sequence is being played.